Restore v0.14 lyrics
The new TextAnnotation version of lyrics in VF4 is too buggy to use: lyrics are spaced beautifully normally, but can easily cause music to overflow the stave, which is impossible to use.

Temporarily restore the old version of lyrics.  (All the new routines are still around (in note.ts and vfShims.ts) but are not being used.)

This will allow development of music21j to continue until VF5 is out and 5.1 can implement a stave overflow patch.